Combine GoToNode and GoToShelf
