+++ src/main/java/frc/robot/controls/GameControllerDriverConfig.java
@@ -2,24 +2,24 @@ package frc.robot.controls;
 
 import java.util.function.DoubleSupplier;
 
+import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.math.geometry.Rotation2d;
+import edu.wpi.first.math.geometry.Transform2d;
+import edu.wpi.first.math.geometry.Translation2d;
 import edu.wpi.first.wpilibj.DriverStation;
 import edu.wpi.first.wpilibj.DriverStation.Alliance;
 import edu.wpi.first.wpilibj.shuffleboard.ShuffleboardTab;
 import edu.wpi.first.wpilibj2.command.InstantCommand;
 import frc.robot.commands.BalanceCommand;
-import frc.robot.commands.GoToNode;
-import frc.robot.commands.GoToNodePID;
-import frc.robot.commands.GoToShelf;
-import frc.robot.commands.GoToShelfPID;
+import frc.robot.commands.GoToPose;
 import frc.robot.commands.SetFormationX;
 import frc.robot.constants.OIConstants;
+import frc.robot.constants.VisionConstants;
 import frc.robot.subsystems.Drivetrain;
 import frc.robot.util.Functions;
 import lib.controllers.GameController;
 import lib.controllers.GameController.Axis;
 import lib.controllers.GameController.Button;
-import lib.controllers.GameController.DPad;
 
 /**
  * Driver controls for the generic game controller.
@@ -41,14 +41,30 @@ public class GameControllerDriverConfig extends BaseDriverConfig {
     kDriver.get(Button.X).onTrue(new SetFormationX(super.getDrivetrain()));
 
     // Moves to the selected scoring position using Path Planner
-    kDriver.get(Button.LB).whileTrue(new GoToNode(m_operator, getDrivetrain(), super.getIntakeOffset()));
+    kDriver.get(Button.LB).whileTrue(new GoToPose(()->getNodePose(), getDrivetrain()));
 
     // Moves to the shelf using Path Planner
-    kDriver.get(Button.RB).whileTrue(new GoToShelf(getDrivetrain()));
+    kDriver.get(Button.RB).whileTrue(new GoToPose(()->(DriverStation.getAlliance()==Alliance.Blue?VisionConstants.kBlueShelfAlignPose:VisionConstants.kRedShelfAlignPose), getDrivetrain()));
     
     kDriver.get(Button.B).whileTrue(new BalanceCommand(super.getDrivetrain()));
 
     kDriver.get(Button.A).onTrue(new InstantCommand(() -> getDrivetrain().resetModulesToAbsolute()));
+  }
+
+  public Pose2d getNodePose(){
+    // get the desired score pose
+    Pose2d scorePose = m_operator.getSelectedNode().scorePose;
+
+    // get a y offset from the supplier (currently we use the intake to offset scoring)
+    double yOffset = getIntakeOffset().getAsDouble();
+
+    // modify pose by offsets
+    scorePose = scorePose.plus(new Transform2d(
+      new Translation2d(0, -0.07 + yOffset), 
+      new Rotation2d(0)
+    ));
+
+    return scorePose;
   }
   
   @Override
